Refactoring initial solution
