+++ packages/core/src/RenderingEngine/Viewport.ts
@@ -17,6 +17,7 @@ import type {
   ICamera,
   ActorEntry,
   IRenderingEngine,
+  IVolumeInput,
   ViewportInputOptions,
   Point2,
   Point3,
@@ -27,6 +28,7 @@ import type {
 import type { ViewportInput, IViewport } from '../types/IViewport';
 import type { vtkSlabCamera } from './vtkClasses/vtkSlabCamera';
 import { getConfiguration } from '../init';
+import { createVolumeActor } from './helpers';
 
 /**
  * An object representing a single viewport, which is a camera
@@ -116,6 +118,107 @@ class Viewport implements IViewport {
 
   static get useCustomRenderingPipeline(): boolean {
     return false;
+  }
+
+  public async isValidVolumeInputArray(
+    volumeInputArray: Array<IVolumeInput>
+  ): Promise<boolean> {
+    return true;
+  }
+
+  public canUse16BitTexture() {
+    return false;
+  }
+
+  protected fixSlabThickness(slabThickness: number) {
+    if (slabThickness) {
+      return slabThickness;
+    } else {
+      slabThickness = 1.0;
+    }
+  }
+  /**
+   * Creates and adds volume actors for all volumes defined in the `volumeInputArray`.
+   * For each entry, if a `callback` is supplied, it will be called with the new volume actor as input.
+   *
+   * @param volumeInputArray - The array of `VolumeInput`s which define the volumes to add.
+   * @param immediate - Whether the `Viewport` should be rendered as soon as volumes are added.
+   */
+  public async addVolumes(
+    volumeInputArray: Array<IVolumeInput>,
+    immediate = false,
+    suppressEvents = false
+  ): Promise<void> {
+    const firstImageVolume = cache.getVolume(volumeInputArray[0].volumeId);
+
+    if (!firstImageVolume) {
+      throw new Error(
+        `imageVolume with id: ${firstImageVolume.volumeId} does not exist`
+      );
+    }
+    const volumeActors = [];
+
+    await this.isValidVolumeInputArray(volumeInputArray);
+
+    // One actor per volume
+    for (let i = 0; i < volumeInputArray.length; i++) {
+      const { volumeId, visibility, actorUID, slabThickness } =
+        volumeInputArray[i];
+
+      const actor = await createVolumeActor(
+        volumeInputArray[i],
+        this.element,
+        this.id,
+        suppressEvents,
+        this.canUse16BitTexture()
+      );
+
+      if (visibility === false) {
+        actor.setVisibility(false);
+      }
+
+      // We cannot use only volumeId since then we cannot have for instance more
+      // than one representation of the same volume (since actors would have the
+      // same name, and we don't allow that) AND We cannot use only any uid, since
+      // we rely on the volume in the cache for mapper. So we prefer actorUID if
+      // it is defined, otherwise we use volumeId for the actor name.
+      const uid = actorUID || volumeId;
+      volumeActors.push({
+        uid,
+        actor,
+        slabThickness: this.fixSlabThickness(slabThickness),
+        // although the actor UID is defined, we need to use the volumeId for the
+        // referenceId, since the actor UID is used to reference the actor in the
+        // viewport, however, the actor is created from its volumeId
+        // and if later we need to grab the referenced volume from cache,
+        // we can use the referenceId to get the volume from the cache
+        referenceId: volumeId,
+      });
+    }
+
+    this.addActors(volumeActors);
+
+    if (immediate) {
+      // render
+      this.render();
+    }
+  }
+
+  /**
+   * It removes the volume actor from the Viewport. If the volume actor is not in
+   * the viewport, it does nothing.
+   * @param actorUIDs - Array of actor UIDs to remove. In case of simple volume it will
+   * be the volume Id, but in case of Segmentation it will be `{volumeId}-{representationType}`
+   * since the same volume can be rendered in multiple representations.
+   * @param immediate - If true, the Viewport will be rendered immediately
+   */
+  public removeVolumeActors(actorUIDs: Array<string>, immediate = false): void {
+    // Todo: This is actually removeActors
+    this.removeActors(actorUIDs);
+
+    if (immediate) {
+      this.render();
+    }
   }
 
   /**
